Make calc_dist_azi test cases robust for angles
In testing that two angles are almost equal, 359.99999
should be considered close to 0.0. Wrap unittest's
assertAlmostEqual so we can have such test cases, and
apply them to the angular tests for taup_geo.

Also move the test cases for taup_geo out of test_tau and
into test_taup_geo.
