+++ obspy/taup/tests/test_taup_geo.py
@@ -10,6 +10,7 @@ from future.builtins import *  # NOQA
 import unittest
 
 from obspy.taup.tau import TauPyModel
+from obspy.taup.taup_geo import calc_dist, calc_dist_azi
 import obspy.geodetics.base as geodetics
 
 
@@ -35,8 +36,97 @@ class TaupGeoTestCase(unittest.TestCase):
             self.assertAlmostEqual(stlon, stlon_path, delta=0.1)
 
 
+class TaupGeoDistTestCase(unittest.TestCase):
+    """
+    Test suite for calc_dist and calc_dist_azi in taup_geo.
+    """
+    def assert_angle_almost_equal(self, first, second, places=7, msg=None,
+                                  delta=None):
+        """
+        Compare two angles (in degrees) for equality
+
+        This method considers numbers close to 359.9999999 to be similar
+        to 0.00000001 and supports the same arguments as assertAlmostEqual
+        """
+        if first > second:
+            difference = (second - first) % 360.0
+        else:
+            difference = (first - second) % 360.0
+        self.assertAlmostEqual(difference, 0.0, places=places, msg=msg,
+                               delta=delta)
+
+    def test_taup_geo_calc_dist(self):
+        """Test for calc_dist"""
+        self.assertAlmostEqual(calc_dist(source_latitude_in_deg=20.0,
+                                         source_longitude_in_deg=33.0,
+                                         receiver_latitude_in_deg=55.0,
+                                         receiver_longitude_in_deg=33.0,
+                                         radius_of_planet_in_km=6371.0,
+                                         flattening_of_planet=0.0), 35.0, 5)
+        self.assertAlmostEqual(calc_dist(source_latitude_in_deg=55.0,
+                                         source_longitude_in_deg=33.0,
+                                         receiver_latitude_in_deg=20.0,
+                                         receiver_longitude_in_deg=33.0,
+                                         radius_of_planet_in_km=6371.0,
+                                         flattening_of_planet=0.0), 35.0, 5)
+        self.assertAlmostEqual(calc_dist(source_latitude_in_deg=-20.0,
+                                         source_longitude_in_deg=33.0,
+                                         receiver_latitude_in_deg=-55.0,
+                                         receiver_longitude_in_deg=33.0,
+                                         radius_of_planet_in_km=6371.0,
+                                         flattening_of_planet=0.0), 35.0, 5)
+        self.assertAlmostEqual(calc_dist(source_latitude_in_deg=-20.0,
+                                         source_longitude_in_deg=33.0,
+                                         receiver_latitude_in_deg=-55.0,
+                                         receiver_longitude_in_deg=33.0,
+                                         radius_of_planet_in_km=6.371,
+                                         flattening_of_planet=0.0), 35.0, 5)
+
+    def test_taup_geo_calc_dist_azi(self):
+        """Test for calc_dist"""
+        dist, azi, backazi = calc_dist_azi(source_latitude_in_deg=20.0,
+                                           source_longitude_in_deg=33.0,
+                                           receiver_latitude_in_deg=55.0,
+                                           receiver_longitude_in_deg=33.0,
+                                           radius_of_planet_in_km=6371.0,
+                                           flattening_of_planet=0.0)
+        self.assertAlmostEqual(dist, 35.0, 5)
+        self.assert_angle_almost_equal(azi, 0.0, 5)
+        self.assert_angle_almost_equal(backazi, 180.0, 5)
+        dist, azi, backazi = calc_dist_azi(source_latitude_in_deg=55.0,
+                                           source_longitude_in_deg=33.0,
+                                           receiver_latitude_in_deg=20.0,
+                                           receiver_longitude_in_deg=33.0,
+                                           radius_of_planet_in_km=6371.0,
+                                           flattening_of_planet=0.0)
+        self.assertAlmostEqual(dist, 35.0, 5)
+        self.assert_angle_almost_equal(azi, 180.0, 5)
+        self.assert_angle_almost_equal(backazi, 0.0, 5)
+        dist, azi, backazi = calc_dist_azi(source_latitude_in_deg=-20.0,
+                                           source_longitude_in_deg=33.0,
+                                           receiver_latitude_in_deg=-55.0,
+                                           receiver_longitude_in_deg=33.0,
+                                           radius_of_planet_in_km=6371.0,
+                                           flattening_of_planet=0.0)
+        self.assertAlmostEqual(dist, 35.0, 5)
+        self.assert_angle_almost_equal(azi, 180.0, 5)
+        self.assert_angle_almost_equal(backazi, 0.0, 5)
+        dist, azi, backazi = calc_dist_azi(source_latitude_in_deg=-20.0,
+                                           source_longitude_in_deg=33.0,
+                                           receiver_latitude_in_deg=-55.0,
+                                           receiver_longitude_in_deg=33.0,
+                                           radius_of_planet_in_km=6.371,
+                                           flattening_of_planet=0.0)
+        self.assertAlmostEqual(dist, 35.0, 5)
+        self.assert_angle_almost_equal(azi, 180.0, 5)
+        self.assert_angle_almost_equal(backazi, 0.0, 5)
+
+
 def suite():
-    return unittest.makeSuite(TaupGeoTestCase, 'test')
+    suite = unittest.TestSuite()
+    suite.addTest(unittest.makeSuite(TaupGeoTestCase, 'test'))
+    suite.addTest(unittest.makeSuite(TaupGeoDistTestCase, 'test'))
+    return suite
 
 
 if __name__ == '__main__':
